Revert "CompressedImageSaver revamp (#23567)"
This reverts commit 3f9bd6bd91659c1e9a92791be4a0ca669cc1df32.
