+++ Cargo.toml
@@ -435,12 +435,7 @@ trace = ["bevy_internal/trace", "dep:tracing"]
 # Basis Universal compressed texture support
 basis-universal = ["bevy_internal/basis-universal"]
 
-# Texture compression asset processor (cross-platform, transcodes to any GPU format at load time)
-compressed_image_saver_universal = [
-  "bevy_internal/compressed_image_saver_universal",
-]
-
-# Texture compression asset processor (BCn for desktop, ASTC for mobile via env var)
+# Enables compressed KTX2 UASTC texture output on the asset processor
 compressed_image_saver = ["bevy_internal/compressed_image_saver"]
 
 # Enables system-level clipboard support.
@@ -2084,18 +2079,6 @@ doc-scrape-examples = true
 [package.metadata.example.asset_decompression]
 name = "Asset Decompression"
 description = "Demonstrates loading a compressed asset"
-category = "Assets"
-wasm = false
-
-[[example]]
-name = "compressed_image_saver"
-path = "examples/asset/compressed_image_saver.rs"
-doc-scrape-examples = true
-required-features = ["compressed_image_saver", "asset_processor", "jpeg"]
-
-[package.metadata.example.compressed_image_saver]
-name = "Compressed Image Saver"
-description = "Demonstrates compressing textures and generating mipmaps using CompressedImageSaver"
 category = "Assets"
 wasm = false
