--- assets/textures/GroundSand005/GroundSand005_COL_2K.jpg.meta
@@ -1,16 +1,0 @@
-(
-    meta_format_version: "1.0",
-    asset: Process(
-        processor: "LoadTransformAndSave<ImageLoader, IdentityAssetTransformer<Image>, CompressedImageSaver>",
-        settings: (
-            loader_settings: (
-                format: FromExtension,
-                is_srgb: true,
-                sampler: Default,
-                asset_usage: ("RENDER_WORLD"),
-            ),
-            transformer_settings: (),
-            saver_settings: (),
-        ),
-    ),
-)
