--- assets/textures/GroundSand005/GroundSand005_DISP_2K.jpg.meta
@@ -1,17 +1,0 @@
-(
-    meta_format_version: "1.0",
-    asset: Process(
-        processor: "LoadTransformAndSave<ImageLoader, IdentityAssetTransformer<Image>, CompressedImageSaver>",
-        settings: (
-            loader_settings: (
-                format: FromExtension,
-                is_srgb: false,
-                sampler: Default,
-                asset_usage: ("RENDER_WORLD"),
-                texture_format: Some("r8unorm"),
-            ),
-            transformer_settings: (),
-            saver_settings: (),
-        ),
-    ),
-)
