--- assets/textures/GroundSand005/source.txt
@@ -1,1 +1,0 @@
-https://www.poliigon.com/texture/rippled-wet-sand-texture/6997
