+++ crates/bevy_internal/Cargo.toml
@@ -29,12 +29,7 @@ detailed_trace = ["bevy_ecs/detailed_trace", "bevy_render?/detailed_trace"]
 
 sysinfo_plugin = ["bevy_diagnostic/sysinfo_plugin"]
 
-# Texture compression asset processor (cross-platform, transcodes to any GPU format at load time)
-compressed_image_saver_universal = [
-  "bevy_image/compressed_image_saver_universal",
-]
-
-# Texture compression asset processor (BCn for desktop, ASTC for mobile via env var)
+# Enables compressed KTX2 UASTC texture output on the asset processor
 compressed_image_saver = ["bevy_image/compressed_image_saver"]
 
 # For ktx2 supercompression
